ci: configure Code Climate test coverage reporting
This configures the Code Climate test coverage reporting. It is a bit
messier than I would like, but I can explain.

- The `codeclimate-test-reporter` has been deprecated in favor of this
binary tool which is used for all languages. For ruby, the preferred
method to generate coverage is SimpleCov.
- A `before-build` command must be run. There is some magic that happens
behind the scenes to generate a JSON version of the SimpleCov report
that Code Climate will accept.
- Only one coverage report is needed, so there is some duplication of
the test command, but it is only run once per ruby version, and only
the latest ruby version publishes coverage.